Ported the update_external_state function to a Celery task
